Implements Imgui class
